// src/api/models/settlementListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Admin
 * Allta Admin 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { SettlementListStoreItem } from './settlementListStoreItem';

export interface SettlementListItem {
  /** ID */
  id: string;
  /** 계좌 은행 */
  bankName: string;
  /** 계좌 번호 */
  accountNumber: string;
  /** 예금주 */
  accountHolder: string;
  /** 수수료율 */
  commissionRate: number;
  /** 총 금액 */
  totalAmount: number;
  /**
   * 정산일
   * @nullable
   */
  settledAt?: string | null;
  /** 정산 여부 */
  isSettled: boolean;
  /** 매장 정보 */
  store: SettlementListStoreItem;
}